requirements.txt
flask
flask-login
flask-sqlalchemy
flask-wtf
flask-mail
flask-bcrypt
sqlalchemy
werkzeug
opencv-python-headless
scikit-learn
numpy
pandas
joblib
requests
python-dotenv
pillow
scikit-image
gunicorn
razorpay
psycopg2-binary
ultralytics
google-generativeai
google-genai
httpx
